+++ es/baseapi.py
@@ -1,3 +1,5 @@
+from typing import Dict, List, Optional, Tuple
+
 from elasticsearch import exceptions as es_exceptions
 from es import exceptions
 from six import string_types
@@ -17,25 +19,76 @@ class Type(object):
 def check_closed(f):
     """Decorator that checks if connection/cursor is closed."""
 
-    def g(self, *args, **kwargs):
+    def wrap(self, *args, **kwargs):
         if self.closed:
             raise exceptions.Error(
                 "{klass} already closed".format(klass=self.__class__.__name__)
             )
         return f(self, *args, **kwargs)
 
-    return g
+    return wrap
 
 
 def check_result(f):
     """Decorator that checks if the cursor has results from `execute`."""
 
-    def g(self, *args, **kwargs):
+    def wrap(self, *args, **kwargs):
         if self._results is None:
             raise exceptions.Error("Called before `execute`")
         return f(self, *args, **kwargs)
 
-    return g
+    return wrap
+
+
+def get_type(data_type):
+    type_map = {
+        "text": Type.STRING,
+        "keyword": Type.STRING,
+        "integer": Type.NUMBER,
+        "half_float": Type.NUMBER,
+        "scaled_float": Type.NUMBER,
+        "geo_point": Type.STRING,
+        # TODO get a solution for nested type
+        "nested": Type.STRING,
+        "object": Type.STRING,
+        "date": Type.DATETIME,
+        "datetime": Type.DATETIME,
+        "short": Type.NUMBER,
+        "long": Type.NUMBER,
+        "float": Type.NUMBER,
+        "double": Type.NUMBER,
+        "bytes": Type.NUMBER,
+        "boolean": Type.BOOLEAN,
+        "ip": Type.STRING,
+        "interval_minute_to_second": Type.STRING,
+        "interval_hour_to_second": Type.STRING,
+        "interval_hour_to_minute": Type.STRING,
+        "interval_day_to_second": Type.STRING,
+        "interval_day_to_minute": Type.STRING,
+        "interval_day_to_hour": Type.STRING,
+        "interval_year_to_month": Type.STRING,
+        "interval_second": Type.STRING,
+        "interval_minute": Type.STRING,
+        "interval_day": Type.STRING,
+        "interval_month": Type.STRING,
+        "interval_year": Type.STRING,
+    }
+    return type_map[data_type.lower()]
+
+
+def get_description_from_columns(columns: List[Dict[str, str]]) -> List[Tuple]:
+    return [
+        (
+            column.get("name") if "alias" not in column else column.get("alias"),
+            get_type(column.get("type")),
+            None,  # [display_size]
+            None,  # [internal_size]
+            None,  # [precision]
+            None,  # [scale]
+            True,  # [null_ok]
+        )
+        for column in columns
+    ]
 
 
 class BaseConnection(object):
@@ -123,16 +176,16 @@ class BaseCursor(object):
     @property
     @check_result
     @check_closed
-    def rowcount(self):
+    def rowcount(self) -> int:
         return len(self._results)
 
     @check_closed
-    def close(self):
+    def close(self) -> None:
         """Close the cursor."""
         self.closed = True
 
     @check_closed
-    def execute(self, operation, parameters=None):
+    def execute(self, operation, parameters=None) -> "BaseCursor":
         raise NotImplementedError  # pragma: no cover
 
     @check_closed
@@ -143,7 +196,7 @@ class BaseCursor(object):
 
     @check_result
     @check_closed
-    def fetchone(self):
+    def fetchone(self) -> Optional[Tuple[str]]:
         """
         Fetch the next row of a query result set, returning a single sequence,
         or `None` when no more data is available.
@@ -155,7 +208,7 @@ class BaseCursor(object):
 
     @check_result
     @check_closed
-    def fetchmany(self, size=None):
+    def fetchmany(self, size: int = None) -> List[Tuple[str]]:
         """
         Fetch the next set of rows of a query result, returning a sequence of
         sequences (e.g. a list of tuples). An empty sequence is returned when
@@ -167,7 +220,7 @@ class BaseCursor(object):
 
     @check_result
     @check_closed
-    def fetchall(self):
+    def fetchall(self) -> List[Tuple[str]]:
         """
         Fetch all (remaining) rows of a query result, returning them as a
         sequence of sequences (e.g. a list of tuples). Note that the cursor's
@@ -202,7 +255,7 @@ class BaseCursor(object):
         # remove dummy schema from queries
         return query.replace(f'FROM "{DEFAULT_SCHEMA}".', "FROM ")
 
-    def elastic_query(self, query: str, csv=False):
+    def elastic_query(self, query: str):
         """
         Request an http SQL query to elasticsearch
         """
@@ -210,10 +263,7 @@ class BaseCursor(object):
         # Sanitize query
         query = self.sanitize_query(query)
         payload = {"query": query, "fetch_size": self.fetch_size}
-        if csv:
-            path = f"/{self.sql_path}/?format=csv"
-        else:
-            path = f"/{self.sql_path}/"
+        path = f"/{self.sql_path}/"
         try:
             resp = self.es.transport.perform_request("POST", path, body=payload)
         except es_exceptions.ConnectionError as e:
